Adjusting file structure for readability and editing readme to reflect changes
