+++ Constants.py
@@ -3,7 +3,6 @@ B_FIELD = 3.12E-5  # T
 E_CHARGE = 1.6022E-19  # Coulombs
 E_MASS = 9.10938356E-31  # kg
 C = 2.998E8  # m / s
-PI = 3.141592653589793  # unit-less
 EPSILON_0 = 8.854187817E-12  # F / m
 B_FACTOR = E_CHARGE / (2 * PI * E_MASS)  # Used to calculate gyro frequency
 TYPE_ABBREVIATION = {'QuasiParabolic': 'QP',
