+++ Tracer.py
@@ -103,6 +103,18 @@ def diagonal_dirs(inputs):
     return np.array([varied_parameter, dp_dx, d2pdx2])
 
 
+class SystemState:
+    def __init__(
+            self, field: Field,
+            atmosphere: Atmosphere,
+            operating_frequency: float,
+            is_extraordinary_ray: bool
+    ):
+        self.field, self.atmosphere = field, atmosphere
+        self.operating_frequency = operating_frequency
+        self.is_extraordinary_ray = is_extraordinary_ray
+
+
 class Tracer:
     def __init__(
             self,
