+++ atmospheres/__init__.py
@@ -1,0 +1,4 @@
+from .base import BaseAtmosphere
+from .implementations import *
+
+__all__ = [s for s in dir() if not s.startswith('_')]
