+++ main.py
@@ -46,11 +46,11 @@ if __name__ == "__main__":
     t0 = time.time()
     basic_tracer.trace(h=10, high_ray=True)
     print(time.time() - t0)
+    atmosphere.visualize()
     basic_tracer.trace(h=10, high_ray=True)
     fig, ax = basic_tracer.visualize(show=False)
     basic_tracer.trace(h=10, high_ray=False)
     basic_tracer.visualize(fig=fig, ax=ax)
-    raise RuntimeError("Done")
     basic_tracer.visualize(show_history=True)
     basic_tracer.trace(h=10, high_ray=False)
     basic_tracer.visualize(show_history=True)
